Move the registry file back
