+++ typescript/infra/config/registry.ts
@@ -1,7 +1,11 @@
 import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
 
-import { ChainAddresses, MergedRegistry } from '@hyperlane-xyz/registry';
+import {
+  ChainAddresses,
+  MergedRegistry,
+  PartialRegistry,
+} from '@hyperlane-xyz/registry';
 import { FileSystemRegistry } from '@hyperlane-xyz/registry/fs';
 import {
   ChainMap,
@@ -12,13 +16,14 @@ import {
 } from '@hyperlane-xyz/sdk';
 import { assert, objFilter, rootLogger } from '@hyperlane-xyz/utils';
 
-import type { DeployEnvironment } from '../../src/config/environment.js';
-import { supportedChainNames as mainnet3Chains } from '../environments/mainnet3/supportedChainNames.js';
+import type { DeployEnvironment } from '../src/config/environment.js';
+
+import { supportedChainNames as mainnet3Chains } from './environments/mainnet3/supportedChainNames.js';
 import {
   testChainMetadata,
   testChainNames as testChains,
-} from '../environments/test/chains.js';
-import { supportedChainNames as testnet4Chains } from '../environments/testnet4/supportedChainNames.js';
+} from './environments/test/chains.js';
+import { supportedChainNames as testnet4Chains } from './environments/testnet4/supportedChainNames.js';
 
 const DEFAULT_REGISTRY_URI = join(
   dirname(fileURLToPath(import.meta.url)),
@@ -109,4 +114,25 @@ export function getMainnetAddresses(): ChainMap<ChainAddresses> {
 
 export function getTestnetAddresses(): ChainMap<ChainAddresses> {
   return getEnvAddresses('testnet4');
+}
+
+let registryWithOverrides: MergedRegistry;
+
+export function getRegistryWithOverrides(
+  chainMetadataOverrides: ChainMap<Partial<ChainMetadata>>,
+): MergedRegistry {
+  if (registryWithOverrides) {
+    return registryWithOverrides;
+  }
+
+  const baseRegistry = getRegistry();
+
+  const overrideRegistry = new PartialRegistry({
+    chainMetadata: chainMetadataOverrides,
+  });
+
+  registryWithOverrides = new MergedRegistry({
+    registries: [baseRegistry, overrideRegistry],
+  });
+  return registryWithOverrides;
 }
